Backport fix for bug 1398 to stable branch.  ifSpeed has been converted to a bigint and code that uses this has been converted to use longs instead of ints.

+++ etc/create.sql
@@ -285,7 +285,7 @@ create table snmpInterface (
 	snmpIfDescr		varchar(256),
 	snmpIfType		integer,
 	snmpIfName		varchar(32),
-	snmpIfSpeed		integer,
+	snmpIfSpeed		bigint,
 	snmpIfAdminStatus	integer,
 	snmpIfOperStatus	integer,
 	snmpIfAlias		varchar(256),
