+++ src/web/src/org/opennms/web/response/ResponseTimeModel.java
@@ -317,8 +317,8 @@ public class ResponseTimeModel extends Object {
             if ((intfMap.get("ipaddr") != null) && !((String) intfMap.get("ipaddr")).equals("0.0.0.0")) {
                 parenString += (String) intfMap.get("ipaddr");
             }
-            if ((intfMap.get("snmpifspeed") != null) && (Integer.parseInt((String) intfMap.get("snmpifspeed")) != 0)) {
-                String speed = Util.getHumanReadableIfSpeed(Integer.parseInt((String) intfMap.get("snmpifspeed")));
+            if ((intfMap.get("snmpifspeed") != null) && (Long.parseLong((String) intfMap.get("snmpifspeed")) != 0)) {
+                String speed = Util.getHumanReadableIfSpeed(Long.parseLong((String) intfMap.get("snmpifspeed")));
                 parenString += ((parenString.equals("")) ? speed : (", " + speed));
             }
